+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 This changelog follows the specifications detailed in: [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
+## Unreleased
+
+### Fixed
+
+* Accept `depends` as an alias for `depends_on` in symbol dependencies.
+
+### Changed
+
+* Warn on unrecognized symbol-spec keys.
+
 ## Version 0.0.2 -- Released 2026-05-08
 
 ### Added